opencode_projects/audio-name/audio-renamer/requirements.txt
mutagen>=1.45.0
PyYAML>=5.4.0